+++ test/korma/test/db.clj
@@ -2,6 +2,45 @@
   (:use clojure.test
         korma.db))
 
+
+(def db-config-with-defaults
+  {:classname "org.h2.Driver"
+   :subprotocol "h2"
+   :subname "mem:db_connectivity_test_db"
+   :user "bob"
+   :password "password"})
+
+(def db-config-with-options-set
+  {:classname "org.h2.Driver"
+   :subprotocol "h2"
+   :subname "mem:db_connectivity_test_db"
+   :excess-timeout 99
+   :idle-timeout 88
+   :minimum-pool-size 5
+   :maximum-pool-size 20})
+
+(deftest connection-pooling-default-test
+  (let [pool (connection-pool db-config-with-defaults)
+        datasource (:datasource pool)]
+    (is (= "org.h2.Driver" (.getDriverClass datasource)))
+    (is (= "jdbc:h2:mem:db_connectivity_test_db" (.getJdbcUrl datasource)))
+    (is (= "bob" (.getUser datasource)))
+    (is (= "password" (.getPassword datasource)))
+
+    (is (= 1800 (.getMaxIdleTimeExcessConnections datasource)))
+    (is (= 10800 (.getMaxIdleTime datasource)))
+    (is (= 3 (.getMinPoolSize datasource)))
+    (is (= 15 (.getMaxPoolSize datasource)))))
+
+(deftest connection-pooling-test
+  (let [pool (connection-pool db-config-with-options-set)
+        datasource (:datasource pool)]
+    (is (= 99 (.getMaxIdleTimeExcessConnections datasource)))
+    (is (= 88 (.getMaxIdleTime datasource)))
+    (is (= 5 (.getMinPoolSize datasource)))
+    (is (= 20 (.getMaxPoolSize datasource)))))
+
 (deftest spec-with-missing-keys-returns-itself
   (defdb valid {:datasource :from-app-server})
   (is (= {:datasource :from-app-server} (get-connection valid))))
+