+++ HISTORY.md
@@ -1,3 +1,4 @@
+*   Set Min/Max Connection Pool Size from db spec (Nick Zalabak)
 *   Merge defaults instead of overwriting them (Jim Crossley)
 *   DB specs can reference existing datasources or JNDI references (Jim Crossley)
 *   Added `between` predicate (Charles Duffy)